# apps/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: AI Centralized Minutes API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
...
  /action-items:
    get:
      summary: List action items with filters
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: projectId
          schema: { type: string }
        - in: query
          name: meetingId
          schema: { type: string }
        - in: query
          name: ownerUserId
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [OPEN, TODO, IN_PROGRESS, BLOCKED, DONE, CANCELLED] }
        - in: query
          name: overdueOnly
          schema: { type: boolean }
        - in: query
          name: dueFrom
          schema: { type: string, format: date-time }
        - in: query
          name: dueTo
          schema: { type: string, format: date-time }
      responses:
        '200': { description: OK }
        '400': { description: Invalid query }
  /action-items/{id}:
    get:
      summary: Get action item detail with status history
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Action item not found }
    patch:
      summary: Update action item quick fields
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string }
                dueDate: { type: string, format: date-time }
                priority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
                ownerDisplayName: { type: string }
      responses:
        '200': { description: Updated }
        '400': { description: Invalid payload }
        '403': { description: Forbidden }
        '404': { description: Action item not found }
  /action-items/{id}/reassign:
    post:
      summary: Reassign action item owner
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ownerUserId]
              properties:
                ownerUserId: { type: string }
                note: { type: string }
      responses:
        '200': { description: Reassigned }
        '400': { description: Invalid payload }
        '404': { description: Action item or owner not found }
  /action-items/{id}/status:
    post:
      summary: Change action item status and write history
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [OPEN, TODO, IN_PROGRESS, BLOCKED, DONE, CANCELLED] }
                note: { type: string }
      responses:
        '200': { description: Updated }
        '400': { description: Invalid payload or transition }
        '403': { description: Forbidden }
        '404': { description: Action item not found }
  /ask-ai:
    post:
      summary: Ask AI grounded on approved minutes and action tracking data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string }
                projectId: { type: string }
                meetingId: { type: string }
                model: { type: string }
                includeRetrievalDebug: { type: boolean, description: Admin-only retrieval trace output }
      responses:
        '200':
          description: Grounded answer with citations and used evidence
          content:
            application/json:
              schema:
                type: object
                required: [answer, confidence, citations, usedEvidence]
                properties:
                  answer: { type: string }
                  confidence: { type: string, enum: [low, medium, high] }
                  uncertainties:
                    type: array
                    items: { type: string }
                  citations:
                    type: array
                    items:
                      type: object
                      properties:
                        chunkId: { type: string }
                        sourceType: { type: string }
                        sourceRowId: { type: string, nullable: true }
                        projectId: { type: string }
                        meetingId: { type: string }
                        meetingTitle: { type: string }
                        meetingDate: { type: string }
                        minuteVersionId: { type: string }
                        minuteApprovedAt: { type: string }
                        snippet: { type: string }
                        hybridScore: { type: number }
                  usedEvidence:
                    type: array
                    items:
                      type: object
                      properties:
                        chunkId: { type: string }
                        sourceType: { type: string }
                        sourceRowId: { type: string, nullable: true }
                        projectId: { type: string }
                        meetingId: { type: string }
                        meetingTitle: { type: string }
                        meetingDate: { type: string }
                        minuteVersionId: { type: string }
                        minuteApprovedAt: { type: string }
                        snippet: { type: string }
                        vectorScore: { type: number }
                        lexicalScore: { type: number }
                        sourceBoost: { type: number }
                        recencyBoost: { type: number }
                        hybridScore: { type: number }
                  usedMeetingIds:
                    type: array
                    items: { type: string }
                  usedActionItemIds:
                    type: array
                    items: { type: string }
                  usedDecisionIds:
                    type: array
                    items: { type: string }
                  retrievalDebug:
                    type: object
                    nullable: true
        '400': { description: Invalid payload }
        '502': { description: Ask-AI failed }
  /ask-ai/logs:
    get:
      summary: List Ask-AI query logs (admin)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: projectId
          schema: { type: string }
        - in: query
          name: meetingId
          schema: { type: string }
        - in: query
          name: userId
          schema: { type: string }
      responses:
        '200': { description: Paginated Ask-AI logs }
        '400': { description: Invalid query }
        '403': { description: Forbidden }
...